portal/requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
psycopg2-binary==2.9.10
kubernetes==33.1.0